requirements.txt
streamlit==1.39.0
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
Pillow==10.4.0
